+++ src/index.ts
@@ -1,3 +1,4 @@
 export * from "./types";
 export * from "./erc20";
 export * from "./accounts";
+export * from "./aave";