+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Next version
 
+- Add function to get all AAVE v3 addresses
+
 # 0.0.2-0
 
 - Initial version with ERC20 and Mangrove accounts